+++ moderator/admin.py
@@ -1,57 +1,27 @@
 from django.contrib import admin
-from django.contrib.admin import SimpleListFilter
-from django.contrib.comments.models import Comment
 from django.contrib.comments.admin import CommentsAdmin as DjangoCommentsAdmin
+from django.contrib.comments.models import Comment
 from django.core.urlresolvers import reverse
-from django.forms.models import BaseInlineFormSet
 from django.template import defaultfilters
-from moderator import constants
-from moderator.models import CannedReply, ClassifiedComment, CommentReply
-from moderator import utils
-
-
-class ClassificationListFilter(SimpleListFilter):
-    title = "classification"
-    parameter_name = "classification"
-
-    def lookups(self, request, model_admin):
-        return constants.CLASS_CHOICES
-
-    def queryset(self, request, queryset):
-        """
-        Returns queryset filtered on categories and primary_category.
-        """
-        if self.value():
-            return queryset.filter(classifiedcomment__cls=self.value())
+from moderator import models, utils
 
 
 class CommentReplyInline(admin.StackedInline):
-    extra = 0
+    extra = 1
     exclude = ['reply_comment', ]
-    model = CommentReply
+    model = models.CommentReply
     fk_name = 'replied_to_comment'
 
 
-class ClassifiedCommentInlineFormSet(BaseInlineFormSet):
-    def save_existing(self, form, instance, commit=True):
-        utils.classify_comment(instance.comment, instance.cls)
-
-
-class ClassifiedCommentInline(admin.StackedInline):
-    formset = ClassifiedCommentInlineFormSet
-    extra = 0
-    model = ClassifiedComment
-
-    def has_add_permission(self, request):
-        return False
-
-
 class CommentAdmin(DjangoCommentsAdmin):
-    list_display = ('comment_text', 'content', 'user', 'submit_date',
-                    'classification', 'moderator_replied',)
-    list_filter = ('submit_date', ClassificationListFilter)
+    list_display = (
+        'comment_text',
+        'content',
+        '_user',
+        'submit_date',
+        'moderator_replied',
+    )
     inlines = [
-        ClassifiedCommentInline,
         CommentReplyInline,
     ]
 
@@ -74,19 +44,66 @@ class CommentAdmin(DjangoCommentsAdmin):
     comment_text.short_description = 'Comment'
     comment_text.allow_tags = True
 
-    def classification(self, obj):
-        try:
-            cls = obj.classifiedcomment_set.get().cls
-        except ClassifiedComment.DoesNotExist:
-            cls = utils.classify_comment(obj).cls
-        return cls.title()
-        return obj.classifiedcomment_set.get().cls.title()
+    def _user(self, obj):
+        url = reverse('admin:auth_user_change', args=(obj.user.id,))
 
+        return '<a href="%s">%s</a>' % (url, obj.user)
+    _user.allow_tags = True
+    _user.short_description = 'User'
+
+
+class CommentProxyAdmin(CommentAdmin):
     def queryset(self, request):
-        qs = super(CommentAdmin, self).queryset(request)
-        return qs.filter(reply_comment_set__isnull=True)
+        qs = super(CommentProxyAdmin, self).queryset(request)
+        return qs.filter(classifiedcomment__cls=self.cls, reply_comment_set__isnull=True)
+
+    def mark_spam(self, modeladmin, request, queryset):
+        for comment in queryset:
+            utils.classify_comment(comment, cls='spam')
+        self.message_user(request, "%s comment(s) successfully marked as spam." % queryset.count())
+    mark_spam.short_description = "Mark selected comments as spam"
+
+    def mark_ham(self, modeladmin, request, queryset):
+        for comment in queryset:
+            utils.classify_comment(comment, cls='ham')
+        self.message_user(request, "%s comment(s) successfully marked as ham." % queryset.count())
+    mark_ham.short_description = "Mark selected comments as ham"
+
+    def get_actions(self, request):
+        actions = {}
+        for action in self.actions:
+            actions[action] = (
+                getattr(self, action),
+                action,
+                getattr(self, action).short_description,
+            )
+        return actions
 
 
-admin.site.register(CannedReply)
+class HamCommentAdmin(CommentProxyAdmin):
+    cls = 'ham'
+    actions = ['mark_spam', ]
+
+
+class ReportedCommentAdmin(CommentProxyAdmin):
+    cls = 'reported'
+    actions = ['mark_ham', 'mark_spam', ]
+
+
+class SpamCommentAdmin(CommentProxyAdmin):
+    cls = 'spam'
+    actions = ['mark_ham', ]
+
+
+class UnsureCommentAdmin(CommentProxyAdmin):
+    cls = 'unsure'
+    actions = ['mark_ham', 'mark_spam', ]
+
+
+admin.site.register(models.CannedReply)
 admin.site.unregister(Comment)
 admin.site.register(Comment, CommentAdmin)
+admin.site.register(models.HamComment, HamCommentAdmin)
+admin.site.register(models.ReportedComment, ReportedCommentAdmin)
+admin.site.register(models.SpamComment, SpamCommentAdmin)
+admin.site.register(models.UnsureComment, UnsureCommentAdmin)