+++ moderator/utils.py
@@ -3,11 +3,16 @@ from django.contrib.contenttypes.models import ContentType
 from moderator.classifier import classifier
 from moderator.constants import DEFAULT_CONFIG
 from secretballot.models import Vote
+from moderator import models
 
 
 def train(comment, is_spam):
     classifier.train(comment.comment, is_spam)
     classifier.store()
+
+
+def clear():
+    classifier.bayes.clear()
 
 
 def get_class(comment):
@@ -39,21 +44,19 @@ def classify_comment(comment, cls=None):
 
     If no class is provided a lookup is done to see if the comment has been
     reported by users as abusive. If indicated as abusive class is set
-    as 'reported', with no training occuring.
+    as 'reported', with spam training occuring and the comment being removed.
 
     If a comment is not reported as abusive by users and without a class being
     provided a class is determined using Baysian inference. In this case no
-    training occurs to prevent circular training.
+    training occurs to prevent self learning.
 
     Returns a newly created or updated ClassifiedComment object.
     As a side effect also sets is_removed field of comment based on class.
     """
-    from moderator.models import ClassifiedComment
-
     if cls not in ['spam', 'ham', 'unsure', 'reported', None]:
         raise Exception("Unrecognized classifications.")
 
-    classified_comment, created = ClassifiedComment.objects.get_or_create(
+    classified_comment, created = models.ClassifiedComment.objects.get_or_create(
         comment=comment
     )
 
@@ -79,6 +82,7 @@ def classify_comment(comment, cls=None):
         return classified_comment
 
     if cls == 'reported' and classified_comment.cls != 'reported':
+        train(comment, is_spam=True)
         comment.is_removed = True
         comment.save()
         classified_comment.cls = cls
@@ -86,15 +90,13 @@ def classify_comment(comment, cls=None):
         return classified_comment
 
     if cls is None:
+        comment_content_type = ContentType.objects.get_for_model(comment)
+        moderator_settings = getattr(settings, 'MODERATOR', DEFAULT_CONFIG)
         if Vote.objects.filter(
-            content_type=ContentType.objects.get_for_model(comment),
+            content_type=comment_content_type,
             object_id=comment.id,
             vote=-1
-        ).count() >= getattr(
-            settings,
-            'MODERATOR',
-            DEFAULT_CONFIG
-        )['ABUSE_CUTOFF']:
+        ).count() >= moderator_settings['ABUSE_CUTOFF']:
             cls = 'reported'
             comment.is_removed = True
             comment.save()